refactor(gui): move and rename files related to menu bar and footer
