+++ api-editor/gui/src/features/menuBar/GenerateAdapters.tsx
@@ -15,10 +15,10 @@ import {
     VStack,
 } from '@chakra-ui/react';
 import React, { useRef, useState } from 'react';
-import { useAppSelector } from '../app/hooks';
-import { selectAnnotationStore } from '../features/annotations/annotationSlice';
-import { AnnotatedPythonPackageBuilder } from '../features/annotatedPackageData/model/AnnotatedPythonPackageBuilder';
-import { selectRawPythonPackage } from '../features/packageData/apiSlice';
+import { useAppSelector } from '../../app/hooks';
+import { selectAnnotationStore } from '../annotations/annotationSlice';
+import { AnnotatedPythonPackageBuilder } from '../annotatedPackageData/model/AnnotatedPythonPackageBuilder';
+import { selectRawPythonPackage } from '../packageData/apiSlice';
 
 interface GenerateAdaptersProps {
     displayInferErrors: (errors: string[]) => void;
